+++ docs/src/markdown/about/changelog.md
@@ -15,6 +15,8 @@
     available in `coloraide.everything` and will be removed at a future time.
 -   **NEW**: Add new channel aliases: `j` for `jz` in Jzazbz and JzCzhz. Also add `c` for `cz` and `h` for `hz` in
     JzCzhz.
+-   **FIX**: Fix a an issue with the CAM16 model's transformation that prevented good round trip with negative
+    lightness.
 -   **FIX**: Ensure that when `harmony` auto creates a cylindrical space from a rectangular space that it checks
     achromatic status in the original color space.
 -   **FIX**: ∆E HCT should use colorfulness, not chroma, in the calculation.